+++ index.js
@@ -1,5 +1,4 @@
 const fs = require('fs');
-const sampleSize = require('lodash.samplesize');
 
 // Load the words into memory:
 const getWords = (fileName) => {
@@ -10,25 +9,6 @@ const OBJECTS = getWords('objects.txt');
 const PREDICATES = getWords('predicates.txt');
 const TEAMS = getWords('teams.txt');
 
-const sample = (words) => {
-  return sampleSize(words, 10);
-}
-
-const pairs = (firstWords, secondWords) => {
-  if(firstWords.length !== secondWords.length) {
-    console.error("Word pair collection lengths must match.");
-    return null; 
-  }
-  
-  const pairedWords = firstWords.map(
-    (firstWord, index) => (`${firstWord}-${secondWords[index]}`)
-  );
-  
-  return pairedWords;
-}
-
-exports.wordPairs = () => {return pairs(sample(PREDICATES), sample(OBJECTS))};
-exports.objects = () => {return sample(OBJECTS)};
-exports.predicates = () => {return sample(PREDICATES)};
-exports.teamPairs = () => {return pairs(sample(PREDICATES), sample(TEAMS))};
-exports.teams = () => {return sample(TEAMS)};
+exports.objects = () => {return OBJECTS};
+exports.predicates = () => {return PREDICATES};
+exports.teams = () => {return TEAMS};
